RPC code split into JSONRPC, NIOJSONRPC and BitcoinRPC #269
JSONRPC (internal, no deps), BitcoinRPC (external, depends on bitcoin and JSON RPC) and NIOJSONRPC (internal, depends on JSONRPC and SwiftNIO only)
