+++ Package.swift
@@ -27,122 +27,66 @@ let package = Package(
         .package(url: "https://github.com/apple/swift-nio-extras.git", from: "1.0.0"),
     ],
     targets: [
-        .target(
-            name: "Bitcoin",
-            dependencies: [
-                "BitcoinTransport",
-                "BitcoinBlockchain",
-                "BitcoinWallet",
-                "BitcoinBase",
-                "BitcoinCrypto"],
+        // Exposed libraries
+        .target(name: "Bitcoin",
+            dependencies: ["BitcoinRPC", "BitcoinTransport", "BitcoinBlockchain", "BitcoinWallet", "BitcoinBase", "BitcoinCrypto"],
             path: "src/bitcoin"),
+        .target(name: "BitcoinRPC", dependencies: ["BitcoinTransport", "BitcoinBlockchain", "JSONRPC"], path: "src/bitcoin-rpc"),
         .target(
             name: "BitcoinTransport",
-            dependencies: [
-                "BitcoinBlockchain",
-                "BitcoinBase",
-                "BitcoinCrypto",
+            dependencies: ["BitcoinBlockchain", "BitcoinBase", "BitcoinCrypto",
                 .product(name: "AsyncAlgorithms", package: "swift-async-algorithms")],
             path: "src/bitcoin-transport"),
         .target(
             name: "BitcoinBlockchain",
-            dependencies: [
-                "BitcoinBase",
-                "BitcoinCrypto",
+            dependencies: ["BitcoinBase", "BitcoinCrypto",
                 .product(name: "AsyncAlgorithms", package: "swift-async-algorithms")],
             path: "src/bitcoin-blockchain"),
-        .target(
-            name: "BitcoinWallet",
-            dependencies: [
-                "BitcoinBase",
-                "BitcoinCrypto"],
-            path: "src/bitcoin-wallet"),
-        .target(
-            name: "BitcoinBase",
-            dependencies: ["BitcoinCrypto"],
-            path: "src/bitcoin-base"),
-        .target(
-            name: "BitcoinCrypto",
-            dependencies: [
-                "ECCHelper",
-                .product(name: "Crypto", package: "swift-crypto")
-            ],
+        .target(name: "BitcoinWallet", dependencies: ["BitcoinBase", "BitcoinCrypto"], path: "src/bitcoin-wallet"),
+        .target(name: "BitcoinBase", dependencies: ["BitcoinCrypto"], path: "src/bitcoin-base"),
+        .target(name: "BitcoinCrypto", dependencies: ["ECCHelper",
+                .product(name: "Crypto", package: "swift-crypto")],
             path: "src/bitcoin-crypto"),
+
+        // Internal libraries
         .target(
-            name: "ECCHelper",
-            dependencies: [.product(name: "LibSECP256k1", package: "secp256k1")],
-            path: "src/ecc-helper"),
-        .target(
-            name: "BitcoinRPC",
+            name: "NIOJSONRPC",
             dependencies: [
-                "BitcoinTransport",
-                "BitcoinBlockchain",
+                "JSONRPC",
                 .product(name: "NIOCore", package: "swift-nio"),
                 .product(name: "NIOFoundationCompat", package: "swift-nio")],
-            path: "src/bitcoin-rpc"),
-        .testTarget(
-            name: "BitcoinBaseTests",
-            dependencies: [
-                "BitcoinBase"],
-            path: "test/bitcoin-base",
-            resources: [
-                .copy("data")
-            ]),
-        .testTarget(
-            name: "BitcoinCryptoTests",
-            dependencies: [
-                "BitcoinCrypto"],
-            path: "test/bitcoin-crypto"),
-        .testTarget(
-            name: "BitcoinWalletTests",
-            dependencies: [
-                "BitcoinWallet"],
-            path: "test/bitcoin-wallet"),
-        .testTarget(
-            name: "BitcoinTransportTests",
-            dependencies: [
-                "BitcoinTransport"],
-            path: "test/bitcoin-transport"),
-        .testTarget(
-            name: "BitcoinRPCTests",
-            dependencies: [
-                "BitcoinRPC"],
-            path: "test/bitcoin-rpc"),
-        .testTarget(
-            name: "BitcoinBlockchainTests",
-            dependencies: [
-                "BitcoinBlockchain"],
-            path: "test/bitcoin-blockchain"),
-        .testTarget(
-            name: "BitcoinTests",
-            dependencies: [
-                "Bitcoin"],
-            path: "test/bitcoin"),
+            path: "src/nio-json-rpc"),
+        .target(name: "JSONRPC", path: "src/json-rpc"),
+        .target(name: "ECCHelper", dependencies: [.product(name: "LibSECP256k1", package: "secp256k1")], path: "src/ecc-helper"),
+
+        // Tests
+        .testTarget(name: "BitcoinTests", dependencies: ["Bitcoin"], path: "test/bitcoin"),
+        .testTarget(name: "BitcoinRPCTests", dependencies: ["BitcoinRPC"], path: "test/bitcoin-rpc"),
+        .testTarget(name: "BitcoinTransportTests", dependencies: ["BitcoinTransport"], path: "test/bitcoin-transport"),
+        .testTarget(name: "BitcoinBlockchainTests", dependencies: ["BitcoinBlockchain"], path: "test/bitcoin-blockchain"),
+        .testTarget(name: "BitcoinWalletTests", dependencies: ["BitcoinWallet"], path: "test/bitcoin-wallet"),
+        .testTarget(name: "BitcoinCryptoTests", dependencies: ["BitcoinCrypto"], path: "test/bitcoin-crypto"),
+        .testTarget(name: "BitcoinBaseTests", dependencies: ["BitcoinBase"], path: "test/bitcoin-base",
+            resources: [.copy("data")]),
+
+        // Executables
         .executableTarget(
             name: "BitcoinNode", dependencies: [
-                "BitcoinTransport",
-                "BitcoinBlockchain",
-                "BitcoinBase",
-                "BitcoinCrypto",
-                "BitcoinRPC",
+                "BitcoinRPC", "BitcoinTransport", "BitcoinBlockchain", "BitcoinBase", "BitcoinCrypto", "NIOJSONRPC", "JSONRPC",
                 .product(name: "Logging", package: "swift-log"),
                 .product(name: "AsyncAlgorithms", package: "swift-async-algorithms"),
                 .product(name: "ArgumentParser", package: "swift-argument-parser"),
                 .product(name: "ServiceLifecycle", package: "swift-service-lifecycle"),
-                .product(name: "NIO", package: "swift-nio"),
-                .product(name: "NIOExtras", package: "swift-nio-extras"),
-                ],
+                .product(name: "NIOCore", package: "swift-nio"),
+                .product(name: "NIOPosix", package: "swift-nio"),
+                .product(name: "NIOExtras", package: "swift-nio-extras")],
             path: "src/bitcoin-node"),
         .executableTarget(
             name: "BitcoinUtility", dependencies: [
-                "BitcoinTransport",
-                "BitcoinBlockchain",
-                "BitcoinWallet",
-                "BitcoinBase",
-                "BitcoinCrypto",
-                "BitcoinRPC",
+                "BitcoinRPC", "BitcoinTransport", "BitcoinBlockchain", "BitcoinWallet", "BitcoinBase", "BitcoinCrypto", "NIOJSONRPC", "JSONRPC",
                 .product(name: "ArgumentParser", package: "swift-argument-parser"),
-                .product(name: "NIO", package: "swift-nio")],
+                .product(name: "NIOCore", package: "swift-nio"),
+                .product(name: "NIOPosix", package: "swift-nio")],
             path: "src/bitcoin-utility")
     ]
 )
