--- src/bitcoin-node/NodeNetwork.swift
@@ -1,4 +1,0 @@
-import ArgumentParser
-import BitcoinTransport
-
-extension NodeNetwork: Decodable, ExpressibleByArgument { }
