+++ src/bitcoin-node/services/RPCService.swift
@@ -1,12 +1,16 @@
-import BitcoinTransport
+import Foundation
+import Logging
+import ServiceLifecycle
+import NIOCore
+import NIOPosix
+
+import JSONRPC
+import NIOJSONRPC
 import BitcoinCrypto
 import BitcoinBase
 import BitcoinBlockchain
+import BitcoinTransport
 import BitcoinRPC
-import ServiceLifecycle
-import NIO
-import Logging
-import Foundation
 
 private let logger = Logger(label: "swift-bitcoin.rpc")
