+++ src/bitcoin-rpc/commands/GetBlockchainInfoCommand.swift
@@ -1,4 +1,5 @@
 import Foundation
+import JSONRPC
 import BitcoinBase
 import BitcoinBlockchain
