+++ src/bitcoin-utility/commands/Node/StartP2P.swift
@@ -1,5 +1,5 @@
 import ArgumentParser
-import BitcoinRPC
+import JSONRPC
 import BitcoinTransport
 
 struct StartP2P: AsyncParsableCommand {
