+++ src/json-rpc/JSONErrorCode.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public enum JSONErrorCode: Int, Codable, Sendable {
+    case parseError = -32700
+    case invalidRequest = -32600
+    case methodNotFound = -32601
+    case invalidParams = -32602
+    case internalError = -32603
+    case other = -32000
+}
