+++ src/nio-json-rpc/CodableCodec.swift
@@ -1,0 +1,47 @@
+import Foundation
+import NIOCore
+
+// bytes to codable and back
+public final class CodableCodec<In, Out>: ChannelInboundHandler, ChannelOutboundHandler where In: Decodable, Out: Encodable {
+    public typealias InboundIn = ByteBuffer
+    public typealias InboundOut = In
+    public typealias OutboundIn = Out
+    public typealias OutboundOut = ByteBuffer
+
+    public init() { }
+
+    private let decoder = JSONDecoder()
+    private let encoder = JSONEncoder()
+
+    // inbound
+    public func channelRead(context: ChannelHandlerContext, data: NIOAny) {
+        var buffer = unwrapInboundIn(data)
+        let data = buffer.readData(length: buffer.readableBytes)!
+        do {
+            //print("--> decoding \(String(decoding: data[..<min(data.startIndex + 100, data.endIndex)], as: UTF8.self))")
+            let decodable = try self.decoder.decode(In.self, from: data)
+            // call next handler
+            context.fireChannelRead(wrapInboundOut(decodable))
+        } catch let error as DecodingError {
+            context.fireErrorCaught(CodecError.badJSON(error))
+        } catch {
+            context.fireErrorCaught(error)
+        }
+    }
+
+    // outbound
+    public func write(context: ChannelHandlerContext, data: NIOAny, promise: EventLoopPromise<Void>?) {
+        do {
+            let encodable = self.unwrapOutboundIn(data)
+            let data = try encoder.encode(encodable)
+            //print("<-- encoding \(String(decoding: data, as: UTF8.self))")
+            var buffer = context.channel.allocator.buffer(capacity: data.count)
+            buffer.writeBytes(data)
+            context.write(wrapOutboundOut(buffer), promise: promise)
+        } catch let error as EncodingError {
+            promise?.fail(CodecError.badJSON(error))
+        } catch {
+            promise?.fail(error)
+        }
+    }
+}
