+++ src/nio-json-rpc/CodecError.swift
@@ -1,0 +1,5 @@
+enum CodecError: Error {
+    case badFraming
+    case badJSON(Error)
+    case requestTooLarge
+}
