+++ src/nio-json-rpc/HalfCloseOnTimeout.swift
@@ -1,0 +1,17 @@
+import NIOCore
+
+public final class HalfCloseOnTimeout: ChannelInboundHandler {
+
+    public typealias InboundIn = Any
+
+    public init() { }
+
+    public func userInboundEventTriggered(context: ChannelHandlerContext, event: Any) {
+        if event is IdleStateHandler.IdleStateEvent {
+            // this will trigger ByteToMessageDecoder::decodeLast which is required to
+            // recognize partial frames
+            context.fireUserInboundEventTriggered(ChannelEvent.inputClosed)
+        }
+        context.fireUserInboundEventTriggered(event)
+    }
+}
